+++ Editor/DeucarianBootstrapPackageConstants.cs
@@ -4,14 +4,20 @@ namespace Deucarian.Bootstrap.Editor
     {
         public const string PackageName = "com.deucarian.bootstrap";
         public const string DisplayName = "Deucarian Bootstrap";
-        public const string Version = "0.1.0";
+        public const string Version = "0.1.1";
         public const string MenuPath = "Tools/Deucarian/Bootstrap/Open Bootstrapper";
         public const string PackageInstallerMenuPath = "Tools/Deucarian/Package Installer";
         public const string LegacyPackageInstallerMenuPath = "Deucarian/Package Installer";
+        public const string RegistryCatalogUrl = "https://raw.githubusercontent.com/Deucarian/Package-Registry/main/packages.json";
+        public const string FallbackCatalogRelativePath = "Editor/PackageCatalogFallback.json";
 
         public const string EditorPackageId = "com.deucarian.editor";
         public const string EditorPackageDisplayName = "Deucarian Editor";
         public const string EditorPackageGitUrl = "https://github.com/Deucarian/Editor.git#main";
+
+        public const string LoggingPackageId = "com.deucarian.logging";
+        public const string LoggingPackageDisplayName = "Deucarian Logging";
+        public const string LoggingPackageGitUrl = "https://github.com/Deucarian/Logging.git#main";
 
         public const string PackageInstallerPackageId = "com.deucarian.package-installer";
         public const string PackageInstallerPackageDisplayName = "Deucarian Package Installer";